install: add support for external devices
This adds the support to use external devices. By external I mean
devices not known by RAUC (listed within the system.conf). So we are
able to use RAUC to initialise the system from a 'special' medium e.g. a
usb-stick which isn't part of the 'normal' update system. To signal it
to RAUC I introduced a cmdline flag called 'rauc.slot.external'. If the
flag is present RAUC won't try to find an active 'normal' slot.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>